add details to branch locations to aid switch reconstruction
Mir.Trans decomposes multi-way `SwitchInt` instructions into chains of
nested conditional branches.  We now annotate those branches with (1)
the index of the current case of the switch and (2) the constant value
used in the comparison.  Currently we stuff this information into the
source location, since that gets preserved all the way down to the
instrumentation output.  Given the additional info, we'll be able to
reconstruct the original switch and better report cases that were never
taken.